Move tests to IT
